result/model/requirements.txt
joblib~=1.2.0
numpy~=1.21.6
pandas~=1.3.5
scikit-learn~=1.0.2
torch
torchvision
Pillow
jieba
opencv-python
requests